Redesign StackWalker implementation
Expands J9StackWalkState to directly hold all necessary fields instead
of accessing them from walkThread/ELS. Initializing those fields is done
in walkStackFrames() for a vm thread and walkContinuationStackFrames() for a
continuation. Then commonWalker is called to walk the stack frames.
This is done to both native implementation and DDR to ensure
consistency.

Also adds assertion for a non-null jitFPRegisterStorageBase to ensure
it's not accessed while walking continuation which cannot hold a
FPRegister.

Signed-off-by: Gengchen Tuo <gengchen.tuo@ibm.com>